# docs/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Provider Agent Liquidity & Coordination Platform",
    "description": "Decision-support prototype API. Phase 2 foundation \u2014 business features are stubbed until later phases.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/providers": {
      "get": {
        "tags": [
          "reference-ledger"
        ],
        "summary": "List Providers",
        "operationId": "list_providers_api_v1_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderRef"
                  },
                  "type": "array",
                  "title": "Response List Providers Api V1 Providers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/areas": {
      "get": {
        "tags": [
          "reference-ledger"
        ],
        "summary": "List Areas",
        "operationId": "list_areas_api_v1_areas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/outlets": {
      "get": {
        "tags": [
          "reference-ledger"
        ],
        "summary": "List Outlets",
        "operationId": "list_outlets_api_v1_outlets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutletListItem"
                  },
                  "type": "array",
                  "title": "Response List Outlets Api V1 Outlets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/outlets/{outlet_id}": {
      "get": {
        "tags": [
          "reference-ledger"
        ],
        "summary": "Get Outlet",
        "operationId": "get_outlet_api_v1_outlets__outlet_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outlet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outlet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutletDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outlets/{outlet_id}/dashboard": {
      "get": {
        "tags": [
          "reference-ledger"
        ],
        "summary": "Get Dashboard",
        "operationId": "get_dashboard_api_v1_outlets__outlet_id__dashboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outlet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outlet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outlets/{outlet_id}/transactions": {
      "get": {
        "tags": [
          "reference-ledger"
        ],
        "summary": "List Transactions",
        "operationId": "list_transactions_api_v1_outlets__outlet_id__transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outlet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outlet Id"
            }
          },
          {
            "name": "provider_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProviderCode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Code"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outlets/{outlet_id}/balances/history": {
      "get": {
        "tags": [
          "reference-ledger"
        ],
        "summary": "Balance History",
        "operationId": "balance_history_api_v1_outlets__outlet_id__balances_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outlet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outlet Id"
            }
          },
          {
            "name": "reserve_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReserveType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reserve Type"
            }
          },
          {
            "name": "provider_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProviderCode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Code"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outlets/{outlet_id}/data-quality": {
      "get": {
        "tags": [
          "reference-ledger"
        ],
        "summary": "Current Data Quality",
        "operationId": "current_data_quality_api_v1_outlets__outlet_id__data_quality_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outlet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outlet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQualityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outlets/{outlet_id}/data-quality/history": {
      "get": {
        "tags": [
          "reference-ledger"
        ],
        "summary": "Data Quality History",
        "operationId": "data_quality_history_api_v1_outlets__outlet_id__data_quality_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outlet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Outlet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQualityHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AreaRef": {
        "properties": {
          "area_id": {
            "type": "string",
            "format": "uuid",
            "title": "Area Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "area_id",
          "code",
          "name"
        ],
        "title": "AreaRef"
      },
      "BalanceHistoryItem": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Snapshot Id"
          },
          "reserve_type": {
            "$ref": "#/components/schemas/ReserveType"
          },
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "balance": {
            "type": "string",
            "title": "Balance"
          },
          "currency_code": {
            "type": "string",
            "title": "Currency Code"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "is_conflicted": {
            "type": "boolean",
            "title": "Is Conflicted",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot_id",
          "reserve_type",
          "outlet_id",
          "balance",
          "currency_code",
          "observed_at",
          "received_at",
          "source_kind"
        ],
        "title": "BalanceHistoryItem"
      },
      "BalanceHistoryResponse": {
        "properties": {
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "reserve_type": {
            "$ref": "#/components/schemas/ReserveType"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BalanceHistoryItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outlet_id",
          "reserve_type",
          "items"
        ],
        "title": "BalanceHistoryResponse"
      },
...
      "DashboardResponse": {
        "properties": {
          "outlet": {
            "$ref": "#/components/schemas/OutletSummary"
          },
          "shared_cash": {
            "$ref": "#/components/schemas/SharedCashDashboard"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderDashboardItem"
            },
            "type": "array",
            "title": "Providers"
          },
          "alerts": {
            "items": {},
            "type": "array",
            "title": "Alerts"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outlet",
          "shared_cash",
          "providers",
          "generated_at"
        ],
        "title": "DashboardResponse"
      },
      "DataQualityHistoryResponse": {
        "properties": {
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "phase": {
            "type": "string",
            "pattern": "^foundation$",
            "title": "Phase",
            "default": "foundation"
          },
          "assessments": {
            "items": {
              "$ref": "#/components/schemas/DataQualityItem"
            },
            "type": "array",
            "title": "Assessments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outlet_id",
          "assessments"
        ],
        "title": "DataQualityHistoryResponse"
      },
      "DataQualityItem": {
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/ProviderCode"
          },
          "assessment": {
            "$ref": "#/components/schemas/QualityAssessmentInput"
          },
          "phase": {
            "type": "string",
            "pattern": "^foundation$",
            "title": "Phase",
            "default": "foundation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "assessment"
        ],
        "title": "DataQualityItem"
      },
      "DataQualityResponse": {
        "properties": {
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "phase": {
            "type": "string",
            "pattern": "^foundation$",
            "title": "Phase",
            "default": "foundation"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "Pre-Phase-4 interim estimates derived from ingestion metadata."
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/DataQualityItem"
            },
            "type": "array",
            "title": "Providers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outlet_id",
          "providers"
        ],
        "title": "DataQualityResponse"
      },
...
      "FeedHealthSummary": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "confidence_modifier": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Modifier"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "confidence_modifier"
        ],
        "title": "FeedHealthSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutletDetailResponse": {
        "properties": {
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "synthetic_code": {
            "type": "string",
            "title": "Synthetic Code"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "area": {
            "$ref": "#/components/schemas/AreaRef"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderRef"
            },
            "type": "array",
            "title": "Providers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outlet_id",
          "synthetic_code",
          "display_name",
          "area",
          "providers"
        ],
        "title": "OutletDetailResponse"
      },
      "OutletListItem": {
        "properties": {
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "synthetic_code": {
            "type": "string",
            "title": "Synthetic Code"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "area_name": {
            "type": "string",
            "title": "Area Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outlet_id",
          "synthetic_code",
          "display_name",
          "area_name"
        ],
        "title": "OutletListItem"
      },
      "OutletSummary": {
        "properties": {
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "synthetic_code": {
            "type": "string",
            "title": "Synthetic Code"
          },
          "area": {
            "type": "string",
            "title": "Area"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outlet_id",
          "synthetic_code",
          "area"
        ],
        "title": "OutletSummary"
      },
...
      "ProjectionSummary": {
        "properties": {
          "shortage_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shortage At"
          },
          "confidence_score": {
            "type": "string",
            "title": "Confidence Score"
          },
          "confidence_level": {
            "type": "string",
            "title": "Confidence Level"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confidence_score",
          "confidence_level"
        ],
        "title": "ProjectionSummary"
      },
      "ProviderCode": {
        "type": "string",
        "enum": [
          "bkash",
          "nagad",
          "rocket"
        ],
        "title": "ProviderCode"
      },
      "ProviderDashboardItem": {
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/ProviderSummary"
          },
          "balance": {
            "type": "string",
            "title": "Balance"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "feed_health": {
            "$ref": "#/components/schemas/FeedHealthSummary"
          },
          "projection": {
            "$ref": "#/components/schemas/ProjectionSummary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "balance",
          "observed_at",
          "feed_health",
          "projection"
        ],
        "title": "ProviderDashboardItem"
      },
      "ProviderRef": {
        "properties": {
          "provider_id": {
            "type": "string",
            "format": "uuid",
            "title": "Provider Id"
          },
          "code": {
            "$ref": "#/components/schemas/ProviderCode"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider_id",
          "code",
          "display_name"
        ],
        "title": "ProviderRef"
      },
      "ProviderSummary": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProviderCode"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "display_name"
        ],
        "title": "ProviderSummary"
      },
      "QualityAssessmentInput": {
        "properties": {
          "data_quality_assessment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Quality Assessment Id"
          },
          "simulation_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Simulation Run Id"
          },
          "ingestion_batch_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingestion Batch Id"
          },
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "status": {
            "$ref": "#/components/schemas/FeedHealthStatus"
          },
          "confidence_modifier": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Modifier"
          },
          "sample_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sample Count"
          },
          "latest_source_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Source At"
          },
          "assessed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assessed At"
          },
          "engine_version": {
            "type": "string",
            "title": "Engine Version"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/QualityIssueInput"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "simulation_run_id",
          "outlet_id",
          "status",
          "confidence_modifier",
          "sample_count",
          "assessed_at",
          "engine_version",
          "summary"
        ],
        "title": "QualityAssessmentInput"
      },
...
      "ReserveType": {
        "type": "string",
        "enum": [
          "shared_cash",
          "provider_e_money"
        ],
        "title": "ReserveType"
      },
...
      "SharedCashDashboard": {
        "properties": {
          "balance": {
            "type": "string",
            "title": "Balance"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "default": "BDT"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "projection": {
            "$ref": "#/components/schemas/ProjectionSummary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "balance",
          "observed_at",
          "projection"
        ],
        "title": "SharedCashDashboard"
      },
...
      "TransactionListResponse": {
        "properties": {
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "type": "array",
            "title": "Transactions"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "outlet_id",
          "transactions",
          "total"
        ],
        "title": "TransactionListResponse"
      },
      "TransactionResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "synthetic_transaction_ref": {
            "type": "string",
            "title": "Synthetic Transaction Ref"
          },
          "synthetic_party_ref": {
            "type": "string",
            "title": "Synthetic Party Ref"
          },
          "provider": {
            "$ref": "#/components/schemas/ProviderCode"
          },
          "transaction_type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "currency_code": {
            "type": "string",
            "title": "Currency Code"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "transaction_id",
          "synthetic_transaction_ref",
          "synthetic_party_ref",
          "provider",
          "transaction_type",
          "status",
          "amount",
          "currency_code",
          "occurred_at",
          "received_at"
        ],
        "title": "TransactionResponse"
      },
      "TransactionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "completed",
          "failed",
          "reversed"
        ],
        "title": "TransactionStatus"
      },
      "TransactionType": {
        "type": "string",
        "enum": [
          "cash_in",
          "cash_out",
          "payment",
          "refund",
          "adjustment"
        ],
        "title": "TransactionType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}